fix - Support conditional on class on jre 8
